Updated strcuture
