+++ constants.py
@@ -1,0 +1,9 @@
+
+# Game settings
+RESOLUTION = WIDTH, HEIGHT = 1600, 900
+HORIZONTAL_RESOLUTION = 360
+HALF_VERTICAL_RESOLUTION = int(HORIZONTAL_RESOLUTION * 0.375)
+SCALING_FACTOR = HORIZONTAL_RESOLUTION / 60
+
+MAP_SIZE = 25
+NUM_ENEMIES = 10
